Fix tests
Note that this is a change in behaviour - previously the `--output` was
incorrect and the usage was displayed as though an error but it was
ignored.

It seems clear that the intention was the `output` subcommand,
so I've left that in rather than remove it, or change the `--`-parsing
behaviour to exit with status 0 on a bad argument in order to continue
to allow it. I've also changed it to a single line output instead of the
full `usage` - since this makes it clearer what's wrong in polybar,
where it may occur if typo'd in config, for example.