Optimize listen function
- Use grep first to receive only the wanted lines
- Flush the pending events to call output only once

With multiple occurrences of pulseaudio-control started, I noticed
that on quick changes (i.e. volume or iterating through the sinks),
the cpu usage increased a lot for the action.

pactl is quite noisy, and a single event can generate multiple
matching lines.

This commit reduces the calls to output to one per 100ms.