Fix buffering output delay
Events always come with multiple lines to process.
In order to process it once per event, do the following:

- Call output to display the changed state
- Wait 100ms to discard all the lines of the event
- Call output again to display the state (which could have changed if
  the user did an action within the 100ms window)

Before, we were always waiting 100ms after an event to display the
output. There was only one call to output, but late.

Without waiting for 100ms, we would be calling output too many times
for a single event.

Having two calls to output per 100ms is the best of the two solutions:
 there is responsiveness, and the number of calls is limited.